codice riorganizzato, lsh forest pronto  da integrare nel main tester

+++ experiments/main/main_tester.py
@@ -1,9 +1,6 @@
 import os
 import shutil
 import argparse
-import warnings
-warnings.filterwarnings('ignore')
-import subprocess
 from time import time
 
 import pandas as pd
@@ -19,7 +16,8 @@ from tools.josiestuff.functions import (
     sample_queries
 )
 
-
+# TODO ok argparse and CLI, maybe better a file .py with variables and import them? 
+# TODO integrate LSHForest testing in main tester 
 parser = argparse.ArgumentParser()
 parser.add_argument('--test-name', 
                     type=str, required=True)
@@ -31,23 +29,29 @@ parser.add_argument('-k',
                     help='the K value for the top-K search of JOSIE')
 parser.add_argument('-t', '--tasks', 
                     required=False, nargs='+',
-                    choices=['all', 'createindex', 'samplequeries', 'dbsetup', 'updatequeries', 'query'], 
-                    help='the tasks to do')
+                    choices=['all', 
+                             'j-createindex', 'f-createforest',
+                             'samplequeries', 
+                             'j-dbsetup', 
+                             'j-updatequeries', 
+                             'j-query', 'f-query'
+                             ], 
+                    help='the tasks to do. \
+                        The prefix "j-" refers to JOSIE tasks, the "f-" to the LSH Forest ones.')
 parser.add_argument('-d', '--dbname', 
                     required=False, default='userdb',
                     help='the PostgreSQL database where will be uploaded the data used by JOSIE. It must be already running on the machine')
-parser.add_argument('-l', '--tables-limit', 
-                    required=False, type=int, default=100000,
-                    help='number of tables to effectivly load for processing from sloth.latest_snapshot_tables')
-
 parser.add_argument('--num-table-sampled', 
                     type=int, required=False, default=100,
                     help='the number of tables that will be sampled from the collections and that will be used as query id for JOSIE (the actual number) \
                         may be less than the specified one due to thresholds tables parameter')
-parser.add_argument('--table-statistics-file', required = False, type=str, help='an absoluth path to an exisisting file containing table statistics used for querying')
-parser.add_argument('--query-file', required=False, type=str, help='an absolute path to an existing file containing the queries which will be used for JOSIE tests')
-
-parser.add_argument('--use-scala', required=False, action='store_true', help='instead of use pure Python implementation of the program, use a Scala version for creating inverted index and integer sets.')
+parser.add_argument('--forest-index', 
+                    required=False, type=str, 
+                    help='the location of the LSH Forest index file that will be used for querying. If it does not exist, \
+                        a new index will be created at that location.')
+parser.add_argument('--query-file', 
+                    required=False, type=str, 
+                    help='an absolute path to an existing file containing the queries which will be used for JOSIE tests')
 parser.add_argument('--small', 
                     required=False, action='store_true',
                     help='works on small collection versions (only for testing)')
@@ -62,12 +66,10 @@ mode =              args.mode
 tasks =             args.tasks if args.tasks else []
 k =                 args.k
 user_dbname =       args.dbname
-tables_limit =      args.tables_limit
-use_scala =         args.use_scala
 small =             args.small
 nsamples =          args.num_table_sampled
 
-# should be a cli argument
+# TODO set thresholds as a CLI parameter or somethig else?
 tables_thresholds = {
     'min_rows':     0,
     'min_columns':  0,
@@ -77,22 +79,19 @@ tables_thresholds = {
     'max_area':     999999,
 }
 
-ALL =               'all' in tasks
-INVERTED_IDX =      'createindex' in tasks
-RAW_TOKENS =        'createrawtokens' in tasks
-SAMPLE_QUERIES =    'samplequeries' in tasks
-DBSETUP =           'dbsetup' in tasks
-UPDATE_QUERY_TABLE = 'updatequeries' in tasks
-JOSIE_TEST =        'josietest' in tasks
+ALL =                   'all' in tasks
+J_CREATE_INVIDX =       'j-createindex' in tasks
+F_CREATE_FOREST =       'f-createforest' in tasks
+SAMPLE_QUERIES =        'samplequeries' in tasks
+J_DBSETUP =             'j-dbsetup' in tasks
+UPDATE_QUERY_TABLE =    'j-updatequeries' in tasks
+JOSIE_TEST =            'j-query' in tasks
+LSHF_QUERY =            'f-query' in tasks
+
+any_task = bool(args.tasks)
+
 CLEAN =             args.clean
 
-# original JSONL and SLOTH results files
-original_turl_train_tables_jsonl_file  =    defpath.data_path.wikitables + '/original_turl_train_tables.jsonl'
-original_sloth_results_csv_file =           defpath.data_path.wikitables + '/original_sloth_results.csv'
-
-# Scala JAR file used for indexing
-scala_jar_indexing_path =                   defpath.root_project_path + '/tools/josiestuff/scala/indexing.jar'
-java_path  =                                '/usr/lib/jvm/java-11-openjdk-amd64/bin/java'
 
 # output files
 ROOT_TEST_DIR =             defpath.data_path.base + f'/josie-tests/{test_name}'
@@ -104,8 +103,6 @@ results_dir =               ROOT_TEST_DIR + '/results'
 
 # statistics stuff
 statistics_dir =            ROOT_TEST_DIR  + '/statistics'
-tables_stat_file =          statistics_dir + '/tables.csv'              if not args.table_statistics_file else args.table_statistics_file
-# columns_stat_file =         statistics_dir + '/columns.csv'
 runtime_stat_file =         statistics_dir + '/runtime.csv'     
 db_stat_file =              statistics_dir + '/db.csv'
 
@@ -113,56 +110,40 @@ runtime_metrics = []
 
 
 ############# SET UP #############
-if os.path.exists(ROOT_TEST_DIR):
-    if input(f'Directory {ROOT_TEST_DIR} already exists: delete it (old data will be lost)? (yes/no) ') in ('y', 'yes'):
-        shutil.rmtree(ROOT_TEST_DIR)
-        
-if not os.path.exists(ROOT_TEST_DIR): 
-    print(f'Creating test directory {ROOT_TEST_DIR}...')
-    os.makedirs(ROOT_TEST_DIR)
-    print(f'Creating test statistics directory {statistics_dir}...')
-    os.makedirs(statistics_dir)
-    print(f'Creating results statistics directory {results_dir}...')
-    os.makedirs(results_dir)
+if any_task:
+    if os.path.exists(ROOT_TEST_DIR):
+        if input(f'Directory {ROOT_TEST_DIR} already exists: delete it (old data will be lost)? (yes/no) ') in ('y', 'yes'):
+            shutil.rmtree(ROOT_TEST_DIR)
+
+    if not os.path.exists(ROOT_TEST_DIR): 
+        print(f'Creating test directory {ROOT_TEST_DIR}...')
+        os.makedirs(ROOT_TEST_DIR)
+        print(f'Creating test statistics directory {statistics_dir}...')
+        os.makedirs(statistics_dir)
+        print(f'Creating results statistics directory {results_dir}...')
+        os.makedirs(results_dir)
 
 ############# DATA PREPARATION #############
-if ALL or INVERTED_IDX:    
+if ALL or J_CREATE_INVIDX:    
     start = time()
-    if not use_scala:
-        create_index(
-            mode,
-            original_sloth_results_csv_file,
-            integer_set_file,
-            inverted_list_file,
-            thresholds=tables_thresholds,
-            tables_limit=tables_limit,
-            small=small
-        )
-    else:
-        print("SCALA VERSION: creating inverted list and integer sets...")
-        subprocess.call(args=[java_path, "-jar", scala_jar_indexing_path, 
-                             mode, 
-                             original_sloth_results_csv_file, 
-                             integer_set_file, 
-                             inverted_list_file, 
-                             tables_limit])
-        print("Completed.")
+    
+    josiedb = JosieDB(dbname=user_dbname, table_prefix=test_name)
+    josiedb.open()
+    josiedb.create_tables()
+    josiedb.close()
+
+    create_index(mode, tables_thresholds, small, test_name)
 
     runtime_metrics.append(('create-invidx-intsets', round(time() - start, 5), get_current_time()))
 
 ############# SAMPLING TEST VALUES FOR JOSIE ##############
 if ALL or SAMPLE_QUERIES:
     start = time()
-    sample_queries(
-        query_file,
-        nsamples,
-        small,
-        tables_thresholds
-    )
+    sample_queries(query_file, nsamples, small, tables_thresholds)
     runtime_metrics.append(('sampling-queries', round(time() - start, 5), get_current_time()))
 
-################### DATABASE OPERATIONS ####################
-if ALL or DBSETUP or UPDATE_QUERY_TABLE:
+################### INSERTING QUERIES INTO POSTGRESQL DATABASE ####################
+if ALL or J_DBSETUP or UPDATE_QUERY_TABLE:
     # reading the IDs for queries
     sampled_ids = get_query_ids_from_query_file(query_file)
         
@@ -172,11 +153,7 @@ if ALL or DBSETUP or UPDATE_QUERY_TABLE:
     josiedb = JosieDB(dbname=user_dbname, table_prefix=test_name)
     josiedb.open()
 
-    if ALL or DBSETUP:
-        josiedb.drop_tables()
-        josiedb.create_tables()
-        josiedb.insert_data_into_inverted_list_table(inverted_list_file)
-        josiedb.insert_data_into_sets_table(integer_set_file)
+    if ALL or J_DBSETUP:
         josiedb.insert_data_into_query_table(sampled_ids)
         josiedb.create_sets_index()
         josiedb.create_inverted_list_index()
@@ -185,10 +162,9 @@ if ALL or DBSETUP or UPDATE_QUERY_TABLE:
         josiedb.insert_data_into_query_table(sampled_ids)
 
     # database statistics
-    dbstat = josiedb.get_statistics()
-    dbstat[0]['test-name'] = test_name
-    pd.DataFrame(dbstat).to_csv(db_stat_file, index=False)
+    pd.DataFrame(josiedb.get_statistics()).to_csv(db_stat_file, index=False)
     josiedb.close()
+    
     runtime_metrics.append(('josie-db-operations', round(time() - start, 5), get_current_time()))
 
 ################## RUNNING JOSIE ##################
@@ -197,11 +173,15 @@ if ALL or JOSIE_TEST:
     josie_test(josie_dbname=user_dbname, test_name=test_name, results_directory=results_dir, k=k)
     runtime_metrics.append(('josie-test', round(time() - start, 5), get_current_time()))
 
-with open(runtime_stat_file, 'a') as rfw:
-    for task in runtime_metrics:
-        rfw.write(f"{task[0]},{task[1]},{task[2]}\n")
+if any_task:
+    with open(runtime_stat_file, 'a') as rfw:
+        for task in runtime_metrics:
+            rfw.write(f"{task[0]},{task[1]},{task[2]}\n")
+    print('All tasks have been completed.')
+
 
 if CLEAN:
+    print('Cleaning directories and database...')
     if os.path.exists(integer_set_file):
         shutil.rmtree(integer_set_file)
     if os.path.exists(inverted_list_file):
@@ -211,7 +191,6 @@ if CLEAN:
     josiedb.open()
     josiedb.drop_tables(all=True)
     josiedb.close()
-
-print('All tasks have been completed.')
+    print('Cleaning completed.')
